Add Audit trail support

+++ apps/retrospective/models/sprint_member.go
@@ -15,11 +15,11 @@ type SprintMember struct {
 	Sprint             Sprint
 	SprintID           uint `gorm:"not null"`
 	Member             userModels.User
-	MemberID           uint `gorm:"not null"`
-	AllocationPercent  uint `gorm:"not null;default:100"`
-	ExpectationPercent uint `gorm:"not null;default:100"`
+	MemberID           uint    `gorm:"not null"`
+	AllocationPercent  float64 `gorm:"not null;default:100"`
+	ExpectationPercent float64 `gorm:"not null;default:100"`
 	Tasks              []SprintMemberTask
-	Vacations          uint                 `gorm:"not null;default:0"`
+	Vacations          float64              `gorm:"not null;default:0"`
 	Rating             retrospective.Rating `gorm:"default:2; not null"`
 	Comment            string               `gorm:"type:text"`
 }
@@ -29,7 +29,7 @@ func (sprintMember *SprintMember) BeforeSave(db *gorm.DB) (err error) {
 	//ToDo: Investigate failing association during SMT save
 	// Vacations should not be more than sprint length
 	//ToDo: Add vacation support - Working Days calculation
-	if sprintMember.Sprint.EndDate != nil && sprintMember.Sprint.EndDate.Sub(*sprintMember.Sprint.StartDate).Hours()/24 < float64(sprintMember.Vacations) {
+	if sprintMember.Sprint.EndDate != nil && sprintMember.Sprint.EndDate.Sub(*sprintMember.Sprint.StartDate).Hours()/24 < sprintMember.Vacations {
 		err = errors.New("vacations cannot be more than sprint length")
 		return err
 	}
